Created HaywardThingHandler super class
