+++ bundles/org.openhab.binding.haywardomnilogic/src/main/java/org/openhab/binding/haywardomnilogic/internal/handler/HaywardChlorinatorHandler.java
@@ -13,16 +13,8 @@
 package org.openhab.binding.haywardomnilogic.internal.handler;
 
 import org.eclipse.jdt.annotation.NonNullByDefault;
-import org.eclipse.jdt.annotation.Nullable;
-import org.eclipse.smarthome.core.library.types.DecimalType;
-import org.eclipse.smarthome.core.thing.Channel;
-import org.eclipse.smarthome.core.thing.ChannelUID;
 import org.eclipse.smarthome.core.thing.Thing;
-import org.eclipse.smarthome.core.thing.ThingStatus;
-import org.eclipse.smarthome.core.thing.binding.BaseThingHandler;
-import org.eclipse.smarthome.core.types.Command;
-import org.eclipse.smarthome.core.types.RefreshType;
-import org.openhab.binding.haywardomnilogic.internal.HaywardBindingConstants;
+import org.openhab.binding.haywardomnilogic.internal.hayward.HaywardThingHandler;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -32,41 +24,10 @@ import org.slf4j.LoggerFactory;
  * @author Matt Myers - Initial Contribution
  */
 @NonNullByDefault
-public class HaywardChlorinatorHandler extends BaseThingHandler {
+public class HaywardChlorinatorHandler extends HaywardThingHandler {
     private final Logger logger = LoggerFactory.getLogger(HaywardChlorinatorHandler.class);
 
     public HaywardChlorinatorHandler(Thing thing) {
         super(thing);
-    }
-
-    @Override
-    public void initialize() {
-        updateStatus(ThingStatus.ONLINE);
-    }
-
-    @Override
-    public void handleCommand(ChannelUID channelUID, Command command) {
-        if (!(command instanceof RefreshType)) {
-            String systemID = getThing().getProperties().get(HaywardBindingConstants.PROPERTY_SYSTEM_ID);
-            String poolID = getThing().getProperties().get(HaywardBindingConstants.PROPERTY_BOWID);
-            HaywardBridgeHandler haywardOmniLogixBridgeHandler = getHaywardOmniLogixBridgeHandler();
-            if (haywardOmniLogixBridgeHandler != null) {
-                haywardOmniLogixBridgeHandler.haywardCommand(channelUID, command, systemID, poolID);
-            }
-
-        }
-    }
-
-    public void updateChlorinatorData(String systemID, String channelID, String data) {
-        Channel chan = getThing().getChannel(channelID);
-        if (chan != null) {
-            updateState(chan.getUID(), new DecimalType(data));
-            logger.trace("Updated Hayward Chlorinator {} {} to: {}", systemID, channelID, data);
-        }
-    }
-
-    @SuppressWarnings("null")
-    private @Nullable HaywardBridgeHandler getHaywardOmniLogixBridgeHandler() {
-        return (HaywardBridgeHandler) getBridge().getHandler();
     }
 }
